refactor: add `created_by` field to the `skins` table. update sample data. enforce ownership before skin update and delete.

+++ apps/backend/src/modules/skins/skin.service.ts
@@ -23,13 +23,19 @@ async function create(requester: User, input: SkinInput): Promise<Skin> {
     if (requester.role !== UserRoleValues.ADMIN)
         throw new AuthorizationError("Only admins can create skins");
 
-    const createdSkin = await skinRepository.create(input);
+    const createdSkin = await skinRepository.create(input, requester.id);
     return toDomain(createdSkin);
 }
 
 async function update(requester: User, id: number, input: SkinInput): Promise<void> {
     if (requester.role !== UserRoleValues.ADMIN)
         throw new AuthorizationError("Only admins can update skins");
+
+    const createdBy = await skinRepository.getCreatedBy(id);
+    if (!createdBy) throw new NotFoundError("Skin not found");
+
+    if (createdBy !== requester.id)
+        throw new AuthorizationError("Only skins created by you can be updated");
 
     const updated = await skinRepository.update(id, input);
     if (!updated) throw new NotFoundError("Skin not found");
@@ -38,6 +44,12 @@ async function update(requester: User, id: number, input: SkinInput): Promise<vo
 async function deleteById(requester: User, id: number): Promise<void> {
     if (requester.role !== UserRoleValues.ADMIN)
         throw new AuthorizationError("Only admins can delete skins");
+
+    const createdBy = await skinRepository.getCreatedBy(id);
+    if (!createdBy) throw new NotFoundError("Skin not found");
+
+    if (createdBy !== requester.id)
+        throw new AuthorizationError("Only skins created by you can be deleted");
 
     const deleted = await skinRepository.deleteById(id);
     if (!deleted) throw new NotFoundError("Skin not found");
